Python/requirements.txt
fastapi
uvicorn
sqlalchemy
aiomysql
pydantic-settings
python-dotenv